+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [DesignToolbox] Update `json` RubyGem from v2.13.1 to v2.13.2
 - [DesignSystemToolbox] Update various GitHub Actions workflows dependencies
+- [DesignToolbox] Theme selector icon (Orange-OpenSource/ouds-ios#943)
+- [DesignToolbox] [Library] Tuning of themes and theme selector (Orange-OpenSource/ouds-ios#951)
 
 ## [0.18.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/0.17.0...0.18.0) - 2025-09-05